+++ lib/lino/utilities.rb
@@ -41,5 +41,13 @@ module Lino
         components
       end
     end
+
+    def or_nil(enumerable, key)
+      enumerable.include?(key) ? enumerable[key] : nil
+    end
+
+    def or_nth(enumerable, key, index)
+      enumerable.include?(key) ? enumerable[key] : enumerable[index]
+    end
   end
 end